add save and load model functions
